Super Sync consume method, removed old unsafe implicit syntax
